optimized the list by separating fastlist from modernlist
