+++ modern-list/src/index.ts
@@ -1,1 +1,3 @@
 export { default } from './lib/modern-list';
+export { default as FastList } from './lib/fast-list';
+export { default as ModernList } from './lib/modern-list';
